+++ stories/primer/button_group_stories.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
-class Primer::ButtonGroupComponentStories < ViewComponent::Storybook::Stories
+class Primer::ButtonGroupStories < ViewComponent::Storybook::Stories
   layout "storybook_preview"
 
   story(:button_group) do
+    controls do
+      select(:variant, Primer::ButtonComponent::VARIANT_OPTIONS, :medium)
+    end
+
     content do |c|
       c.button { "Button" }
       c.button(scheme: :primary) { "Primary" }